fixed data presentation on website; updated hibernate, replaced with new JPA; fixed database

+++ spatial.election.database/src/main/java/edu/spatial/election/database/dao/ConstituencyDAO.java
@@ -8,10 +8,10 @@ import edu.spatial.election.domain.Constituency;
 
 public interface ConstituencyDAO extends SimpleDAO {
 
-	public Constituency findConstituencyById(long id) throws ConstituencyNotFoundException;
+	public Constituency findConstituencyById(int id) throws ConstituencyNotFoundException;
 	public List<Constituency> findConstituencyByState(String stateName);
 	public void saveConstituency(Constituency constituency);
-	public void deleteConstituency(long id) throws ConstituencyNotFoundException;
+	public void deleteConstituency(int id) throws ConstituencyNotFoundException;
 	public List<Constituency> getConstituencies();
 
 }
