+++ spatial.election.database/src/main/java/edu/spatial/election/database/dao/CountyDAO.java
@@ -10,12 +10,12 @@ import edu.spatial.election.domain.County;
 
 public interface CountyDAO extends SimpleDAO{
 	
-	public County findCountyById(long id) throws CountyNotFoundException;
+	public County findCountyById(int id) throws CountyNotFoundException;
 	public List<County> findCountyByState(String stateName);
 	public List<County> findCountyByDistrict(String districtName);
 	public List<County> getCounties();
-	public List<Constituency> findConstituenciesOfCounty(long countyId) throws ConstituencyNotFoundException;
+	public List<Constituency> findConstituenciesOfCounty(int countyId) throws ConstituencyNotFoundException;
 	public void saveCounty(County county);
-	public void deleteCounty(long id) throws CountyNotFoundException;
+	public void deleteCounty(int id) throws CountyNotFoundException;
 
 }
