+++ spatial.election.database/src/main/java/edu/spatial/election/database/dao/PartyDAO.java
@@ -7,8 +7,8 @@ import edu.spatial.election.domain.Party;
 
 public interface PartyDAO extends SimpleDAO {
 
-	public Party findPartyById(long id) throws PartyNotFoundException;
+	public Party findPartyById(int id) throws PartyNotFoundException;
 	public List<Party> getParties();
 	public void saveParty(Party party);
-	public void deleteParty(long id) throws PartyNotFoundException;
+	public void deleteParty(int id) throws PartyNotFoundException;
 }
